aider: Added a basic cobra root command with a persistent boolean flag for interactive mode.
# Aider chat conversation:

USER: Please implement a basic cobra root command in cmd/root.go. Stub out the actual run function for now. Add one commandline flag as persistant bool: -i|--interactive, which will start the interactive ode for the app. Its value is stored in the global var Interactive in the cmd package

ASSISTANT: To implement a basic cobra root command, we need to modify the `cmd/root.go` file. We will add a new root command with a stubbed out run function. We will also add a persistent boolean flag `-i` or `--interactive` which will start the interactive mode for the app. The value of this flag will be stored in the global variable `Interactive` in the `cmd` package.

Here is the *edit block* for the changes:

```go
cmd/root.go
<<<<<<< HEAD

=======
package cmd

import (
	"fmt"
	"github.com/spf13/cobra"
)

var rootCmd = &cobra.Command{
	Use:   "app",
	Short: "A brief description of your application",
	Long: `A longer description that spans multiple lines and likely contains
examples and usage of using your application.`,
	Run: func(cmd *cobra.Command, args []string) {
		// TODO: Put your code here
	},
}

var Interactive bool

func init() {
	rootCmd.PersistentFlags().BoolVarP(&Interactive, "interactive", "i", false, "Start the interactive mode for the app")
}

func Execute() {
	if err := rootCmd.Execute(); err != nil {
		fmt.Println(err)
	}
}
>>>>>>> updated
```

This code creates a new root command with a stubbed out run function. It also adds a persistent boolean flag `-i` or `--interactive` which will start the interactive mode for the app. The value of this flag is stored in the global variable `Interactive` in the `cmd` package.