aider: Added skeleton play command that checks if the file exists before attempting to play it.
# Aider chat conversation:

USER: Now, please implement a skeleton play command, which takes one argument. This should be the file path to a playable music file, in either mp3, flac, or wav format. The command should output an error if the file can't be found, but not exit the app.

ASSISTANT: To implement the play command, we need to modify the `cmd/play.go` file. We will add a new command named `play` which takes one argument, the file path to a playable music file. We will also add a check to ensure the file exists before attempting to play it. If the file does not exist, an error message will be printed to the console.

Here is the code change:

```go
cmd/play.go
<<<<<<< HEAD
=======
package cmd

import (
	"fmt"
	"os"
	"github.com/spf13/cobra"
)

var playCmd = &cobra.Command{
	Use:   "play [file]",
	Short: "Play a music file",
	Long: `Play a music file. The file must be in either mp3, flac, or wav format.`,
	Args:  cobra.ExactArgs(1),
	Run: func(cmd *cobra.Command, args []string) {
		file := args[0]
		if _, err := os.Stat(file); os.IsNotExist(err) {
			fmt.Printf("File %s does not exist\n", file)
			return
		}
		// TODO: Add code to play the file
	},
}

func init() {
	RootCmd.AddCommand(playCmd)
}
>>>>>>> updated
```

This code adds a new command `play` to the root command. The `play` command takes exactly one argument, which is the file path to the music file. The `Run` function checks if the file exists. If the file does not exist, it prints an error message and returns. If the file exists, it should play the file. The code to play the file is not included in this change and should be added later.