aider: Modified print statements to use "\r" and not create a new line, and added a print statement to clear the line before printing the new playback position.
# Aider chat conversation:

ASSISTANT: To achieve this, we can use the "\r" character to return the cursor to the start of the line before printing the new message. This will overwrite the previous message on the same line. We will also need to clear the line before printing a new message to ensure that shorter messages properly overwrite longer ones. We can do this by printing a line of spaces.

Here are the changes we need to make:

1. Modify the print statement for the user instructions to use "\r" and not create a new line.
2. Modify the print statement for the playback position to use "\r" and not create a new line.
3. Add a print statement to clear the line before printing the new playback position.

Here are the edit blocks for these changes:

```go
cmd/play.go
<<<<<<< HEAD
			fmt.Print("Press [ENTER] to pause/resume, 'q' to stop. ")
=======
			fmt.Print("\rPress [ENTER] to pause/resume, 'q' to stop. ")
>>>>>>> updated
```

```go
cmd/play.go
<<<<<<< HEAD
			fmt.Println(format.SampleRate.D(streamer.Position()).Round(time.Second))
=======
			fmt.Print("\r                                                                 \r")
			fmt.Print(format.SampleRate.D(streamer.Position()).Round(time.Second))
>>>>>>> updated
```